requirements.txt
streamlit
torch
transformers
mysql-connector-python
pandas
textblob
mysql-connector-python
pymysql
sqlalchemy
cryptography
